+++ src/models.py
@@ -26,7 +26,6 @@ from .allennlp_mods.elmo_text_field_embedder import ElmoTextFieldEmbedder, ElmoT
 from .utils.utils import assert_for_log, get_batch_utilization, \
     get_batch_size, get_elmo_mixing_weights
 from .utils import config
-from . import edge_probing
 
 from .tasks.tasks import CCGTaggingTask, ClassificationTask, CoLATask, EdgeProbingTask, GroundedSWTask, \
     GroundedTask, LanguageModelingTask, MTTask, MultiNLIDiagnosticTask, PairClassificationTask, \
@@ -41,7 +40,8 @@ from .modules.modules import SentenceEncoder, BoWSentEncoder, \
     NullPhraseLayer
 
 from .preprocess import parse_task_list_arg, get_tasks
-from .seq2seq_decoder import Seq2SeqDecoder
+from .modules.edge_probing import EdgeClassifierModule
+from .modules.seq2seq_decoder import Seq2SeqDecoder
 
 
 # Elmo stuff
@@ -365,7 +365,7 @@ def build_module(task, model, d_sent, d_emb, vocab, embedder, args):
         hid2tag = build_tagger(task, d_sent, task.num_tags)
         setattr(model, '%s_mdl' % task.name, hid2tag)
     elif isinstance(task, EdgeProbingTask):
-        module = edge_probing.EdgeClassifierModule(task, d_sent, task_params)
+        module = EdgeClassifierModule(task, d_sent, task_params)
         setattr(model, '%s_mdl' % task.name, module)
     elif isinstance(task, (RedditSeq2SeqTask, Wiki103Seq2SeqTask)):
         log.info("using {} attention".format(args.s2s['attention']))
