+++ src/modules/edge_probing.py
@@ -6,8 +6,8 @@ import numpy as np
 import torch.nn.functional as F
 from torch.autograd import Variable
 
-from .tasks.tasks import EdgeProbingTask
-from .modules import modules
+from ..tasks.tasks import EdgeProbingTask
+from .import modules
 
 from allennlp.modules.span_extractors import \
     EndpointSpanExtractor, SelfAttentiveSpanExtractor
